requirements.txt
Django>=1.8,<1.9
djangorestframework==3.4.0
django-queryinspect
